Move RubricLogo to components folder
