+++ lib/components/rubric_logo.dart
