+++ lib/landing.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flow_builder/flow_builder.dart';
 
 import 'package:rubric/components/colors.dart';
+import 'package:rubric/components/rubric_logo.dart';
 import 'package:rubric/onboarding_bottom_sheet.dart';
-import 'package:rubric/rubric_logo.dart';
 
 class Landing extends StatelessWidget {
   final FlowController flowController;
